requirements.txt
Django==4.2
django-dotenv==1.4.2
djangorestframework==3.14.0